remove dockercompose, serve static ui from express, update dockerfile
