+++ Dockerfile
@@ -4,6 +4,10 @@ WORKDIR /usr/src/app
 
 COPY . /usr/src/app/
 
-RUN npm install -g @angular/cli
 RUN npm install
 RUN npm run build
+
+EXPOSE 8085
+
+CMD [ "npm", "run", "start-server" ]
+
