--- docker-compose.yml
@@ -1,19 +1,0 @@
-version: '3.5'
-
-services: 
-    story-points: 
-      container_name: story-points  
-      build: . 
-      volumes:
-        - './:/usr/src/app' 
-      ports: 
-        - '8080:8080'
-      command: > 
-        bash -c "npm install && ng serve --host 0.0.0.0 --port 8080"
-    story-points-server:
-      container_name: server
-      build: ./server
-      ports:
-        - '8999:8999'
-      command: > 
-        bash -c "npm install && node server.js"
