--- server/Dockerfile
@@ -1,12 +1,0 @@
-FROM node:10
-
-WORKDIR /usr/src/app
-
-COPY package*.json ./
-
-RUN npm install
-
-COPY . .
-
-EXPOSE 8999
-CMD [ "node", "server.js" ]
