--- server/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "npm",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nohup node server.js &"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nbpeth/socket-server.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/nbpeth/socket-server/issues"
-  },
-  "homepage": "https://github.com/nbpeth/socket-server#readme",
-  "dependencies": {
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "http": "0.0.0",
-    "ws": "^7.1.2"
-  }
-}
